components: move ctx-resource owners (MusicContext / TextContext / RenderTargets / ShapeMeshes) beside owning systems (#1351)
Per .squad/decisions.md §"app/components parallel audit verdict
(consolidated)", app/components/ is reserved for atomic, queryable
entity-owned data. The four GPU/audio resource-owner structs called
out in #1351 each own a non-trivial destructor, deleted copy, custom
move and a release() that invokes raylib Unload*, so they were
ctx-singleton RAII wrappers misfiled under components/.

Moves:
- app/components/music.h           -> app/systems/audio_resources.h
- app/components/text_resources.h  -> app/systems/text_resources.h
- camera_resources.h split:
    RAII   RenderTargets / camera::ShapeMeshes
                                   -> app/systems/camera_resources.h
    data   ScreenTransform / FloorParams / floor_visuals::*
           stay in app/components/camera_resources.h

Include sites updated (game_loop.cpp, song_playback_system.cpp,
play_session.cpp, ui_render_system.cpp, camera_system.cpp,
game_render_system.cpp, tests). app/components/text.h docstring
points at the new TextContext location.

Verification:
- grep 'Unload' app/components/  ->  no matches
- cmake --build build            ->  zero warnings (-Wall -Wextra -Werror)
- ./build/shapeshifter_tests     ->  3357 assertions / 959 test cases all green
