+++ app/components/text.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // UI/domain-facing text enums used by ECS components.
-// Runtime font resources are owned by TextContext in app/components/text_resources.h.
+// Runtime font resources are owned by TextContext in app/systems/text_resources.h.
 
 // Named constants for the pre-loaded font sizes.
 enum class FontSize : int { Small = 0, Medium = 1, Large = 2 };
