+++ app/game_loop.cpp
@@ -10,7 +10,7 @@
 #include "systems/audio_routing.h"
 #include "entities/settings.h"
 #include "components/rhythm.h"
-#include "components/music.h"
+#include "systems/audio_resources.h"
 #include "components/rendering.h"
 #include "systems/test_player.h"
 #include "components/obstacle.h"
@@ -19,9 +19,10 @@
 #include "systems/runtime_systems.h"
 #include "systems/test_player_session.h"
 #include "systems/input_routing.h"
-#include "components/text_resources.h"
+#include "systems/text_resources.h"
 #include "systems/session_logger_system.h"
 #include "systems/camera_system.h"
+#include "systems/camera_resources.h"
 #include "entities/beat_map.h"
 #include "platform_display.h"
 #include "systems/persistence_policy_system.h"
