+++ app/systems/play_session.cpp
@@ -8,7 +8,7 @@
 #include "../components/rhythm.h"
 #include "../components/song_state.h"
 #include "../components/energy_bar.h"
-#include "../components/music.h"
+#include "audio_resources.h"
 #include "../components/high_score.h"
 #include "high_score_system.h"
 #include "../entities/settings.h"
