+++ tests/test_gpu_resource_lifecycle.cpp
@@ -15,10 +15,11 @@
 #include <type_traits>
 
 #include "components/audio.h"
-#include "components/music.h"
+#include "systems/audio_resources.h"
 #include "components/camera_resources.h"
+#include "systems/camera_resources.h"
 #include "entities/camera_entity.h"  // GameCamera, UICamera
-#include "components/text_resources.h"
+#include "systems/text_resources.h"
 #include "systems/session_logger_system.h"
 
 // ── ShapeMeshes type traits ──────────────────────────────────────────────────
